Move pagination hooks to hooks/pagination
