+++ frontend/src/components/cases/case-table.tsx
@@ -13,7 +13,7 @@ import { columns } from "@/components/cases/case-table-columns"
 import { DataTable, type DataTableToolbarProps } from "@/components/data-table"
 import { TooltipProvider } from "@/components/ui/tooltip"
 import { useToast } from "@/components/ui/use-toast"
-import { useCasesPagination } from "@/hooks/use-cases-pagination"
+import { useCasesPagination } from "@/hooks"
 import { getDisplayName } from "@/lib/auth"
 import { useDeleteCase } from "@/lib/hooks"
 import { useAuth } from "@/providers/auth"
