+++ frontend/src/hooks/index.ts
@@ -1,4 +1,3 @@
 export * from "./canvas"
-export * from "./use-cases-pagination"
+export * from "./pagination"
 export * from "./use-debounce"
-export * from "./use-workflows-pagination"
